+++ src/AngleSharp.Js/Cache/PrototypeCache.cs
