+++ src/AngleSharp.Js/Cache/ReferenceCache.cs
