+++ src/AngleSharp.Js/Proxies/ConsoleInstance.cs
