+++ src/AngleSharp.Js/Proxies/DomConstructorInstance.cs
