+++ src/AngleSharp.Js/Proxies/DomDelegateInstance.cs
