+++ src/AngleSharp.Js/Proxies/DomEventInstance.cs
