+++ src/AngleSharp.Js/Proxies/DomFunctionInstance.cs
