+++ src/AngleSharp.Js/Proxies/DomNodeInstance.cs
