+++ src/AngleSharp.Js/Proxies/DomPrototypeInstance.cs
