--- src/AngleSharp.Js/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="jint" version="2.10.4" targetFramework="portable45-net45+win8+wp8+wpa81" />
-</packages>
